# server/crates/networkd/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlightPortrait network API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://data.flightportrait.com"
    }
  ],
  "paths": {
...
    "/v1/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Roster",
        "description": "Public feeder list. Locations rounded to about 11 km. IPs are not stored. Rate: 120 per 600 s (bucket `stations`). Cache: 30 s edge.",
        "operationId": "stations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "stations": {
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Public id, fp-<10 hex>. Reveals no UUID bits."
                          },
                          "label": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "description": "Operator-set display label."
                          },
                          "coarse_lat": {
                            "type": [
                              "number",
                              "null"
                            ],
                            "description": "Coverage midpoint, rounded ~11 km."
                          },
                          "coarse_lon": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "first_seen": {
                            "type": "string",
                            "description": "ISO 8601 UTC."
                          },
                          "last_seen": {
                            "type": "string",
                            "description": "ISO 8601 UTC."
                          },
                          "online": {
                            "type": "boolean"
                          }
                        },
                        "type": "object",
                        "required": [
                          "id",
                          "first_seen",
                          "last_seen",
                          "online"
                        ]
                      },
                      "type": "array"
                    }
                  },
                  "type": "object",
                  "required": [
                    "stations"
                  ]
                },
                "example": {
                  "stations": [
                    {
                      "id": "fp-a1b2c3d4e5",
                      "coarse_lat": 1.4,
                      "coarse_lon": 103.8,
                      "first_seen": "2026-08-01T00:00:00+00:00",
                      "last_seen": "2026-08-29T12:00:00+00:00",
                      "online": true
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          }
        },
        "x-stability": "stable"
      }
    },
    "/v1/stations/{station_uuid}": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Station",
        "description": "Status for one feeder. The full UUID is the key; the server stores a hash. Unknown and malformed UUIDs both return 404. The UUID travels in the URL: treat the link as the secret it carries. Rate: 60 per 600 s (bucket `station_detail`). Never cached.",
        "operationId": "station",
        "parameters": [
          {
            "name": "station_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Full feeder UUID. The server stores a hash, not the UUID.",
              "title": "Station Uuid"
            },
            "description": "Full feeder UUID. The server stores a hash, not the UUID."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "online": {
                      "type": "boolean"
                    },
                    "connected_since": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "ISO 8601 UTC. Null when offline."
                    },
                    "messages_per_s": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "positions_per_s": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "kbit_s": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "rtt_ms": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "positions_total": {
                      "type": "integer"
                    },
                    "aircraft_seen": {
                      "type": [
                        "integer",
                        "null"
                      ],
                      "description": "Aircraft this station sees right now. Null when unavailable."
                    },
                    "first_seen": {
                      "type": "string"
                    },
                    "last_seen": {
                      "type": "string"
                    },
                    "recent_sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "started_at": {
                            "type": "string"
                          },
                          "ended_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "description": "Null while open."
                          },
                          "peak_messages_per_s": {
                            "type": "number"
                          },
                          "positions_total": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "id",
                    "online",
                    "positions_total",
                    "first_seen",
                    "last_seen",
                    "recent_sessions"
                  ]
                },
                "example": {
                  "id": "fp-a1b2c3d4e5",
                  "online": true,
                  "messages_per_s": 0.426,
                  "positions_per_s": 0.109,
                  "kbit_s": 0.09,
                  "rtt_ms": 16,
                  "positions_total": 204,
                  "aircraft_seen": 3,
                  "first_seen": "2026-08-01T00:00:00+00:00",
                  "last_seen": "2026-08-29T12:00:00+00:00",
                  "recent_sessions": [
                    {
                      "started_at": "2026-08-29T00:00:00+00:00",
                      "peak_messages_per_s": 0.5,
                      "positions_total": 204
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          },
          "404": {
            "description": "Not found or not observed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-stability": "stable"
      }
    },
...
    "/v1/setup/beacon": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Stations waiting on my network",
        "description": "The Stations waiting for setup on the caller's own network (the same public address), newest first, each with the address of its setup page. Empty when there are none. Rate: 120 per 600 s (bucket `beacon_find`). Never cached.",
        "operationId": "setup_beacon_get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "stations": [
                    {
                      "url": "http://192.168.1.23:8654/",
                      "name": "station",
                      "seen_s": 12
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          }
        },
        "x-stability": "map"
      },
      "post": {
        "tags": [
          "Stations"
        ],
        "summary": "Station waiting for setup",
        "description": "Sent by a Station that has not been set up yet: its address on the home network, so the join page on a phone on that network can open its setup page. Held in memory for 10 minutes under a hash of the caller's public address; never stored. Only private IPv4 addresses are accepted. Rate: 30 per 600 s (bucket `beacon`).",
        "operationId": "setup_beacon_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beacon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          },
          "422": {
            "description": "Malformed input.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "x-stability": "map"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Beacon": {
        "properties": {
          "lan": {
            "type": "string",
            "title": "Lan"
          },
          "port": {
            "type": "integer",
            "title": "Port",
            "default": 8654
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "station"
          }
        },
        "type": "object",
        "required": [
          "lan"
        ],
        "title": "Beacon"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}